Fix keystone-api's log openening failures on rkt
When the rkt container runtime is used, stdout and stderr are Unix sockets to journald, which can't be opened with the open() syscall, leading a the failure below:

```
AH00649: could not open transfer log file /proc/self/fd/1.
AH00015: Unable to open logs
```

This patch uses "cat" to directly write to the stdio fds without
reopening them. It works both with Docker and rkt.