Generate Horizon assets offline
This fixes an issue that happens when Horizon is used behind a
load-balancer that doesn't do sticky-sessions. The first time the user
would open Horizon, the reached controller would generate the assets but
the browser might ask another server for these assets. And because
apache2 serves these files as static files, the other server's
controller wouldn't be hit and generate the missing asserts.

It also forces the hashing method to 'hash', which is the current
default for Horizon, that is equivalent to the documented 'content'
method in django-compressor 1.4 (https://goo.gl/PPhhWq).