nova: report VMs' memory usage as part of the nova-compute pods
By moving the VM processes to the nova-compute container's
memory cgroup, Kubernetes can correctly monitor the memory
that Nova Compute and its VMs consume. Therefore, operators
have a better vision of what Nova represents on their nodes,
and we give Kubernetes the opportunity to manage the hosts'
resources appropriately given the actual load, the defined
QoS, etc. For instance, if a Nova instance ends up consuming
more than it is allowed to, or if that Nova instance has a
lesser priority than other workloads, Kubernetes could decide
to kill the pod, thus distributing all VMs across other
available Nova pods using the preStop hook.

We also propagate the memory request set for the Nova pod to
the Nova scheduler, so that the scheduler can make better
decisions, and the hypervisor is shown with the correct amount
of memory available.

Fixes #140