Enforce 'replicas: 1' on glance-api when local storage is used
When local storage is used, running multiple instances of
glance-api causes image data to be scattered across multiple containers
and nodes. Therefore, when nova or the glance CLI wants to download an
image, if the request doesn't hit the instance that has the data (due to
load balancing), no data is returned - generating either an exception or
an another unexpected behavior.